@@ -1,66 +1,25 @@
-import "./App.css";
-import Navbar from "./components/Navbar/Navbar";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import React from "react";
-import Trains from "./components/Trains/Trains";
-import Login from "./components/User/Login";
-import { useEffect } from "react";
-import Logout from "./components/User/Logout";
-import { useDispatch } from "react-redux";
-import { check } from "./actions/user";
-import Cookies from "js-cookie";
-import Profile from "./components/Profile/Profile";
-import Signup from "./components/User/Signup";
-import Admin from "./components/Admin/Admin";
-import NotFound from "./components/NotFound/NotFound";
+import logo from './logo.svg';
+import './App.css';
 
-const App = () => {
-  const token = Cookies.get("token");
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    if (token) {
-      dispatch(check());
-    } else {
-      console.log("you are not logged in");
-    }
-  });
-
+function App() {
   return (
-    <Router>
-      {/* Navbar */}
-      <Navbar />
-
-      <Switch>
-        {/* Train List */}
-        <Route exact path="/">
-          <Trains />
-        </Route>
-
-        <Route exact path="/login">
-          <Login />
-        </Route>
-
-        <Route exact path="/logout">
-          <Logout />
-        </Route>
-
-        <Route exact path="/signup">
-          <Signup />
-        </Route>
-
-        <Route exact path="/admin">
-          <Admin />
-        </Route>
-
-        <Route exact path="/profile">
-          <Profile />
-        </Route>
-
-        <Route component={NotFound} />
-      </Switch>
-    </Router>
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.js</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
+    </div>
   );
-};
+}
 
 export default App;
